Re-adding Countries_and_Codes and FIFA Rankings
Added more naming conventions to countries_and_codes.csv to account for historical names of countries on TransferMarkt. Added a new copy of FIFA Rankings that has the country codes we are using assigned. The countries which dropped out of existence or split into multiple countries have 'No Code' for their country code.